+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Added "configure" button to custom feed header.
+- Remove OAuth client credentials from local storage if they are not valid.
 
 ## [4.16.0] - 2026-01-02